ai_services/ai_design_generation_visualization/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []

[project]
name = "ai-design-generation-visualization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "openai>=1.60.0",
    "openai-agents>=0.9.0",
    "aiohttp>=3.11.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.1.0",
    "aiokafka>=0.12.0",
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.9",
    "httpx>=0.27.0",
]
